增加新手引导功能 全局button组件替换为自定义button组件 不再使用遮挡版屏蔽点击事件 增加攻击和偷取中途退出，下次进入直接进入攻击偷取界面的功能 转盘面板现在可以直接进入岛屿界面和转盘界面

+++ Assets/Scripts/Data/AttackData.cs
@@ -16,5 +16,5 @@ public class AttackData{
     public bool isNewCannonContestScoreAward;  // 暂时未使用
     public int betCount;                  // 暂时未使用
     public AttackTargetData attackTarget;                   // 攻击对象  // AttackTarget 数据结构在上面
-    public long tutorial;                  // 新手教程
+    public int tutorial;                  // 新手教程
 }
