+++ Assets/Scripts/Data/StealData.cs
@@ -9,5 +9,5 @@ public class StealData {
     public TargetData[] targets;                       // 三个目标供玩家猜测  // Target 数据结构在上面
     public TargetData stealTarget;             // 偷取对象
     public bool badLuck;                  // 暂时未使用
-    public long tutorial;                   // 新手教程
+    public int tutorial;                   // 新手教程
 }
